+++ lib/src/common_utils.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/cupertino.dart';
+
+TextEditingController dropdownSearchController = TextEditingController(
+  text: '',
+);
+
+const spaceWidth5 = SizedBox(width: 5);