+++ CHANGELOG.md
@@ -2,5 +2,13 @@
 
 ## [0.0.1-dev.1]
 ### Added
-- Initial dev release of `dropdownpro`.
+- Initial dev release of `dropdownproplus`.
 - Overlay dropdown on text field click.
+
+
+## [0.0.1-dev.2]
+### Added
+- Initial dev release of `dropdownproplus`.
+- Overlay dropdown on text field click.
+- UI modifications and functionality enhancement
+- Pagination part is under development